cac: run py gen at runtime too, pass ctx to tasks
In Firedrake/PyOP2, the kernels compiled from UFL expressions
are a small part of what is needed to run the actual task at
runtime: the rest are data structures initialized in Python.
These same data structures are initialized in order to compile
the kernel, so we re-use the same code path (the generator
code path) for runtime too. In principle, these could be separated,
but it would be a huge re-architecting of PyOP2: it's compile
path would need to be stripped of everything that is not
strictly necessary for compilation.